add apparels and fabrics with update apparel

+++ database/migrations/2022_05_25_222307_create_fabrics_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->nullable();
             $table->integer('qty')->nullable();
-            $table->double('price', 10,8)->nullable();
+            $table->float('price', 12,2)->nullable();
             $table->string('image')->nullable();
             $table->string('slug')->nullable();
             $table->string('status')->default('INACTIVE');
